+++ src/main/java/mul/cam/a/dao/LocalDao.java
@@ -3,6 +3,7 @@ package mul.cam.a.dao;
 import mul.cam.a.dto.LocalDto;
 
 public interface LocalDao {	
+	
 	LocalDto getLocal(int seq);
 
 }
