+++ src/main/java/mul/cam/a/dto/LocalDto.java
@@ -3,8 +3,9 @@ package mul.cam.a.dto;
 import java.io.Serializable;
 
 public class LocalDto implements Serializable {
-	private int seq;		// sequence 글번호
+	private int seq;		// sequence 지역번호
 	private String regionname;		// 지역이름(서울, 경기)
+	private int readcount;
 	
 	public LocalDto() {
 		
@@ -32,10 +33,19 @@ public class LocalDto implements Serializable {
 		this.regionname = regionname;
 	}
 
+	public int getReadcount() {
+		return readcount;
+	}
+
+	public void setReadcount(int readcount) {
+		this.readcount = readcount;
+	}
+
 	@Override
 	public String toString() {
-		return "LocalDto [seq=" + seq + ", regionname=" + regionname + "]";
+		return "LocalDto [seq=" + seq + ", regionname=" + regionname + ", readcount=" + readcount + "]";
 	}
+	
 	
 	
 }
