+++ src/main/java/sqls/Local.xml
@@ -4,8 +4,9 @@
 	
 <mapper namespace="Local.">
 
+<!-- dto안에 LocalDto에 있는거로 내보내(return)겠다 -->
 <select id="getLocal" parameterType="Integer" resultType="mul.cam.a.dto.LocalDto">
-	select seq, regionname
+	select seq, regionname, readcount
 	from region
 	where seq=#{seq}
 </select>
