+++ src/main/webapp/WEB-INF/views/layouts.xml
@@ -8,74 +8,81 @@
 <!-- 맨처음 보이는 로그인 화면이라 main -->
 <definition name="main" template="/WEB-INF/views/mainscreentiles.jsp">
 	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-<!-- 	<put-attribute name="menu" value="/WEB-INF/views/menu.jsp"/> -->
 	<put-attribute name="content" value="/WEB-INF/views/content.jsp"/>
-<!-- 	<put-attribute name="empty" value="/WEB-INF/views/empty.jsp"/> -->
 	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
 </definition>
 
 <!-- 지역별 여행 -->
 <definition name="localtravel" template="/WEB-INF/views/localtravelScreen/localtraveltiles.jsp">
 	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<!-- <put-attribute name="menu" value="/WEB-INF/views/menu.jsp"/>
-	<put-attribute name="content" value="/WEB-INF/views/content.jsp"/> -->
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 경기 -->
-<definition name="gyeonggi" template="/WEB-INF/views/localtravelScreen/region/gyeonggi.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 제주 -->
-<definition name="jeju" template="/WEB-INF/views/localtravelScreen/region/jeju.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 강원 -->
-<definition name="gangwon" template="/WEB-INF/views/localtravelScreen/region/gangwon.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 충북 -->
-<definition name="chungbuk" template="/WEB-INF/views/localtravelScreen/region/chungbuk.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 충남 -->
-<definition name="chungnam" template="/WEB-INF/views/localtravelScreen/region/chungnam.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 경북 -->
-<definition name="gyeongbuk" template="/WEB-INF/views/localtravelScreen/region/gyeongbuk.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 경남 -->
-<definition name="gyeongnam" template="/WEB-INF/views/localtravelScreen/region/gyeongnam.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 전북 -->
-<definition name="jeonbuk" template="/WEB-INF/views/localtravelScreen/region/jeonbuk.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
-	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
-</definition>
-<!-- 전남 -->
-<definition name="jeonnam" template="/WEB-INF/views/localtravelScreen/region/jeonnam.jsp">
-	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
-	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/weather.jsp"/>
 	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
 </definition>
 
+<!-- 경기 -->
+<definition name="gyeonggi" template="/WEB-INF/views/localtravelScreen/localtraveltiles2.jsp">
+	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/gyeonggi/ggw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/gyeonggi/gyeonggi.jsp"/>
+	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
+</definition>
+
+<definition name="ggp1" template="/WEB-INF/views/localtravelScreen/localtraveltiles3.jsp">
+	<put-attribute name="header" value="/WEB-INF/views/header.jsp"/>
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/gyeonggi/ggw.jsp"/>
+	<put-attribute name="placemenu" value="/WEB-INF/views/localtravelScreen/placemenu.jsp"/>
+	<put-attribute name="map" value="/WEB-INF/views/localtravelScreen/map.jsp"/>
+	<put-attribute name="placebbs" value="/WEB-INF/views/localtravelScreen/placebbs.jsp"/>
+	<put-attribute name="footer" value="/WEB-INF/views/footer.jsp"/>
+</definition>
+
+
+<!-- 제주 -->
+<definition name="jeju" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/jeju/jjw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/jeju/jeju.jsp"/>
+</definition>
+
+<!-- 강원 -->
+<definition name="gangwon" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/gangwon/gww.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/gangwon/gangwon.jsp"/>
+</definition>
+
+<!-- 충북 -->
+<definition name="chungbuk" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/chungbuk/cbw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/chungbuk/chungbuk.jsp"/>
+</definition>
+
+<!-- 충남 -->
+<definition name="chungnam" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/chungnam/cnw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/chungnam/chungnam.jsp"/>
+</definition>
+
+<!-- 경북 -->
+<definition name="gyeongbuk" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/gyeongbuk/gbw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/gyeongbuk/gyeongbuk.jsp"/>
+</definition>
+
+<!-- 경남 -->
+<definition name="gyeongnam" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/gyeongnam/gnw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/gyeongnam/gyeongbuk.jsp"/>
+</definition>
+
+<!-- 전북 -->
+<definition name="jeonbuk" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/jeonbuk/jbw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/jeonbuk/jeonbuk.jsp"/>
+</definition>
+
+<!-- 전남 -->
+<definition name="jeonnam" extends="localtravel">
+	<put-attribute name="weather" value="/WEB-INF/views/localtravelScreen/region/jeonnam/jnw.jsp"/>
+	<put-attribute name="eachregion" value="/WEB-INF/views/localtravelScreen/region/jeonnam/jeonnam.jsp"/>
+</definition>
 
 
 <!-- 숙소추천 -->
