+++ src/main/webapp/WEB-INF/views/localtravelScreen/localtraveltiles.jsp
@@ -3,7 +3,9 @@
     pageEncoding="UTF-8"%>
 <%@ taglib prefix="tiles" uri="http://tiles.apache.org/tags-tiles"%>
 
-
+<%-- <%	
+	LocalDto dto = (LocalDto)request.getAttribute("localdto");
+%> --%>
 
 <!DOCTYPE html>
 <html>
@@ -27,82 +29,36 @@
 </style>
 </head>
 <body>
-<%	
-	LocalDto dto = (LocalDto)request.getAttribute("localdto");
-%>
+<div class="container">
 <%-- <%=dto.getSeq() %>
 <%=dto.getRegionname() %> --%>
-
-	<div><tiles:insertAttribute name="header"/></div>
-	<div align="center">가고싶은 여행지를 골라보세요</div>
-	
-<!-- 	해봤는데 안됨
-<div id="tab-1" class="tab-content current hidden-sm-down">
-		<div class="col-12 row pl-3 mx-0 pr-0">
-			<div class="col-lg-4 col-md-6 col-4 pr-3 pl-0 pb-3 country_pannel" id="country_korea">
-				<div class="col-12" 
-				style="padding:unset;
-				background-image: url(../images/서울2.jpg);
-				background-repeat: no-repeat;
-				background-position: center; height:350px; 
-				width:100%; 
-				background-size: cover; 
-				" id="country_korea"> 
-					<div style="position:absolute; top:45%; width:100%" class="nation-title txt-h4 weight-600 text-center f-white" >
-					</div>
-				</div>
-				<div id="country_name_korea" style="position:absolute; top:70%; width:100%" 
-						class="nation-title weight-600 text-center f-white country_name" >
-					<div class="txt-h2 align-text-bottom" >
-						KOREA							
-					</div>
-				</div>
-			</div>
-			<div class="col-lg-4 col-md-6 col-4 pr-3 pl-0 pb-3 country_pannel" id="country_korea">
-				<div class="col-12" 
-				style="padding:unset;
-				background-image: url(../images/서울2.jpg);
-				background-repeat: no-repeat;
-				background-position: center; height:350px; 
-				width:100%; 
-				background-size: cover; 
-				" id="country_korea"> 
-					<div style="position:absolute; top:45%; width:100%" class="nation-title txt-h4 weight-600 text-center f-white" >
-					</div>
-				</div>
-				<div id="country_name_korea" style="position:absolute; top:70%; width:100%" 
-						class="nation-title weight-600 text-center f-white country_name" >
-					<div class="txt-h2 align-text-bottom" >
-						KOREA							
-					</div>
-				</div>
-			</div>
-			<div class="col-lg-4 col-md-6 col-4 pr-3 pl-0 pb-3 country_pannel" id="country_korea">
-				<div class="col-12" 
-				style="padding:unset;
-				background-image: url(../images/서울2.jpg);
-				background-repeat: no-repeat;
-				background-position: center; height:350px; 
-				width:100%; 
-				background-size: cover; 
-				" id="country_korea"> 
-					<div style="position:absolute; top:45%; width:100%" class="nation-title txt-h4 weight-600 text-center f-white" >
-					</div>
-				</div>
-				<div id="country_name_korea" style="position:absolute; top:70%; width:100%" 
-						class="nation-title weight-600 text-center f-white country_name" >
-					<div class="txt-h2 align-text-bottom" >
-						KOREA							
-					</div>
-				</div>
-			</div>
-		</div>
+<div><tiles:insertAttribute name="header"/></div>
+	<div align="center"> 가고싶은 지역 선택
+		<!-- 검색버튼 넣을때 사용 
+		<table style="margin-left: auto; margin-right: auto; margin-top: 3px; margin-bottom: 3px">
+			<tr>
+				<td style="padding-left: 3px">
+					<select class="custom-select" id="choice" name="choice">
+						<option selected>검색</option>
+						<option value="title">제목</option>
+						<option value="content">내용</option>
+						<option value="writer">작성자</option>
+					</select>
+				</td>
+				<td style="padding-left: 5px" class="align-middle">
+					<input type="text" class="form-control" id="search" name="search" onkeyup="enterKeyEvent()" placeholder="검색어" value="<%--search --%>">
+				<td style="padding-left: 5px">
+					<span>
+						<button type="button" class="btn btn-primary" onclick="searchBtn()">검색</button>
+					</span>
+				</td>
+			</tr>
+		</table> 
+		-->
 	</div>
- -->
+	
 	<div class="d-flex justify-content-center" align="center">
-		<!-- <div id="Rpic"> 서울 페이지로
-		<button type="button" onclick="location.href='localtravel2.do'">서울</button>
-		</div> -->
+	
 		<div class="Rpicwrap">
 			<div class="Rpicimage"><img src="images/서울2.jpg" alt="" width="250" height="250"></div>
 			<div class="Rpictext">
@@ -170,12 +126,13 @@
 			</div>
 		</div>
 	</div>
-
-	<div><tiles:insertAttribute name="footer"/></div>	
+	
+<div><tiles:insertAttribute name="footer"/></div>	
 	
 
 
-<%-- <div align="center">
+<%-- table태그로 하던거
+ <div align="center">
 
 <table style="width: 100%; height: 100%; border-color: gray;" >
 
@@ -265,5 +222,6 @@
 
 </div>
  --%>
+ </div>
 </body>
 </html>
