+++ src/main/webapp/WEB-INF/views/localtravelScreen/region/gyeonggi/gyeonggi.jsp
@@ -1,0 +1,39 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+<h1>경기도 페이지임</h1>
+<div class="d-flex justify-content-center">
+		<div id="Rpic"> 
+		<a href="ggp1.do" class="textlink">서울/경기</a>	
+		</div>
+		
+		
+		<div id="Rpic">
+		<button type="button" onclick="location.href='main.do'">제주</button>
+		</div>
+		
+		<div id="Rpic">
+		<button type="button" onclick="location.href='main.do'">강원</button>
+		</div>
+	</div>
+	<div class="d-flex justify-content-center">	
+		<div id="Rpic"> 
+		<button type="button" onclick="location.href='main.do'">충북</button>
+		</div>
+		
+		<div id="Rpic"> 
+		<button type="button" onclick="location.href='main.do'">충남</button>
+		</div>
+		
+		<div id="Rpic">
+		<button type="button" onclick="location.href='main.do'">경북</button>
+		</div>
+	</div>
+</body>
+</html>
