+++ src/main/webapp/WEB-INF/views/localtravelScreen/region/jeonbuk/jbw.jsp
@@ -5,13 +5,10 @@
 <head>
 <meta charset="UTF-8">
 <title>Insert title here</title>
-
 </head>
 <body>
-
 <div align="center">
 <img src="images/메인2.jpg" width="800" height="400">
 </div>
-
 </body>
 </html>
